Modified Detail screen to use primary station info when appropriate. Modified event time function in OccultEvent to use individual station data.

+++ OW_Mobile/UtilityFunctions.swift
@@ -304,7 +304,7 @@ func stationSigmaIcon(_ sigmaIconValue: Int?) -> UIImage
   return sigmaIconImage
 }
 
-func moonAltIcon(_ moonAltIconValue: Int?) -> UIImage
+func moonAltIcon(_ moonAltIconValue: Double?) -> UIImage
 {
   var moonAltIconImage: UIImage
   if moonAltIconValue != nil
